Move helper filt into image folder that is the only one using it
