+++ src/scripts/comic/getLinkListPusher.ts
